Fix pagination issue in ListAll
The ListAll method repeatedly calls listChecks to fetch a page of results, incrementing
the page number each time. To detect whether there is another page of results to follow,
it looks at the Next field of the result object:

for clResp.Next != "" {

However, when the clResp variable is updated inside the loop, it was shadowed, by using
the := form:

clResp, _, err := listChecks(...)

This means that the clResp variable in the enclosing scope was never updated, so
clResp.Next was always non-empty, meaning that sooner or later the code will attempt to
fetch a non-existent page, with the resultant error:

GET https://uptime.com/api/v1/checks/?page=27&page_size=1000 failed: Code=NOT_FOUND
Message=Invalid page.

This commit fixes that issue by changing the := to just =, meaning that clResp is no
longer shadowed.